MagicConfigHandler -> MagicConfigHandlerImpl
