+++ magiclib-malilib-extra/src/main/java/top/hendrixshen/magiclib/impl/malilib/SharedConstants.java
@@ -5,9 +5,10 @@ import lombok.Getter;
 import org.jetbrains.annotations.NotNull;
 
 import top.hendrixshen.magiclib.api.i18n.I18n;
+import top.hendrixshen.magiclib.api.malilib.config.MagicConfigHandler;
 import top.hendrixshen.magiclib.api.malilib.config.MagicConfigManager;
 import top.hendrixshen.magiclib.impl.malilib.config.GlobalConfigManager;
-import top.hendrixshen.magiclib.impl.malilib.config.MagicConfigHandler;
+import top.hendrixshen.magiclib.impl.malilib.config.MagicConfigHandlerImpl;
 import top.hendrixshen.magiclib.util.VersionUtil;
 
 public class SharedConstants {
@@ -23,8 +24,8 @@ public class SharedConstants {
     private static final MagicConfigManager configManager = GlobalConfigManager
             .getConfigManager(top.hendrixshen.magiclib.SharedConstants.getMagiclibIdentifier());
     @Getter
-    private static final MagicConfigHandler configHandler = new MagicConfigHandler(SharedConstants.configManager,
-            1);
+    private static final MagicConfigHandler configHandler = new MagicConfigHandlerImpl(SharedConstants.configManager,
+            2);
 
     public static @NotNull String getTranslatedModVersionType() {
         return VersionUtil.translateVersionType(SharedConstants.modVersion);
